+++ client/views/Dashboard/BarChart/BarChart.jsx
@@ -109,6 +109,7 @@ BarChart = React.createClass({
     	var issues = Issues.find(query);
 
     	var buckets = {};
+        var costs = {};
     	var labels = [];
     	var counts = [];
     	issues.map(function(i){
@@ -118,11 +119,15 @@ BarChart = React.createClass({
                 if(serviceName.length>15) {
                     serviceName = serviceName.substring(0,13)+'...';
                 }
+                if(!costs[serviceName]) {
+                    costs[serviceName] = 0;
+                }
     			if(!buckets[serviceName]) {
     				labels.push(serviceName);
     				buckets[serviceName] = [];
     			}
     			buckets[serviceName].push(i);
+                costs[serviceName] += parseInt(i.costThreshold);
     		}
     	});
     	labels.map(function(serviceName,idx){
@@ -132,7 +137,7 @@ BarChart = React.createClass({
     	return {
     		facility:facility,
     		labels:labels,
-    		set:counts
+    		set:costs//counts
     	}
     },
 
@@ -202,7 +207,7 @@ BarChart = React.createClass({
             <div>
                 <ActionsMenu items={this.getMenu()} icon="eye" />
                 <div className="ibox-title">
-                    <h2>Types of requests {this.state.title}</h2>
+                    <h2>Request breakdown {this.state.title}</h2>
                 </div>
                 <div className="ibox-content">
                     <div style={{margin:"0px 25px 0px 0px"}}>