+++ client/views/Facility/FacilityViewEdit.jsx
@@ -74,17 +74,20 @@ FacilityViewEdit = React.createClass({
 			        	</div>
 			        </div>
 				</CollapseBox>
-				<CollapseBox title="Files">
+				<CollapseBox title="Building Documents">
+					<AutoForm item={facility} schema={schema} form={["documents"]}/>
+				</CollapseBox>
+				<CollapseBox title="Quick Files">
 					<AutoForm item={facility} schema={schema} form={["attachments"]}/>
 				</CollapseBox>
-				<CollapseBox title="Members">
+				<CollapseBox title="Facility personnel">
 			   		<ContactList 
 			   			items={members}
 			   			facility={facility}
 			   			onAdd={team.canInviteMember()?this.addMember.bind(null,{role:"staff"}):null}
 			   		/>
 				</CollapseBox>
-				<CollapseBox title="Suppliers">
+				<CollapseBox title="Facility suppliers">
 			   		<ContactList 
 			   			items={suppliers}
 			   			facility={facility}